CHANGE remove unused code
