+++ core/errors.py
@@ -1,0 +1,5 @@
+
+
+class NotFoundOnDb(Exception):
+    def __str__(self):
+        return 'Object not found on database'
